rename "sagemaker" to "aws_sagemaker" for consistency with "aws_batch"
