+++ rastervision_aws_sagemaker/MANIFEST.in
